src/lex-gen-ai-demo-cdk/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
sagemaker==2.163.0
boto3